requirements.txt
Flask==2.1.2
Flask-Mail==0.9.1
python-dotenv==0.20.0
Werkzeug==2.1.2
